refactor(sct-1091): Refactor component/page folder (#659)
